+++ R/8_convert.R
@@ -1,6 +1,6 @@
 #' Convert simulation output to different formats.
 #' 
-#' For use with other packages compatible with dyno / anndata.
+#' For use with other packages compatible with dyno, anndata, SingleCellExperiment, or Seurat.
 #' 
 #' @param model A dyngen output model for which the experiment has been emulated with [generate_experiment()].
 #' @param store_cellwise_grn Whether or not to also store cellwise GRN information.
@@ -129,23 +129,6 @@ as_dyno <- function(
   }
   
   dataset
-}
-
-#' @rdname convert
-#' @export
-wrap_dataset <- function(
-  model,
-  store_dimred = !is.null(model$simulations$dimred),
-  store_cellwise_grn = !is.null(model$experiment$cellwise_grn),
-  store_rna_velocity = !is.null(model$experiment$rna_velocity)
-) {
-  .Deprecated("as_dyno")
-  as_dyno(
-    model = model, 
-    store_dimred = store_dimred,
-    store_cellwise_grn = store_cellwise_grn,
-    store_rna_velocity = store_rna_velocity
-  )
 }
 
 
@@ -386,4 +369,123 @@ as_seurat <- function(
   }
   
   seurat_obj
+}
+
+convert_progressions_to_milestone_percentages <- function (progressions) {
+  selfs <- progressions %>% filter(.data$from == .data$to) %>% select(.data$cell_id, milestone_id = .data$from) %>% mutate(percentage = 1)
+  progressions <- progressions %>% filter(.data$from != .data$to)
+  from_mls <- tapply(progressions$from, progressions$cell_id, first, default = NA_character_)
+  from_pct <- 1 - tapply(progressions$percentage, progressions$cell_id, sum, default = NA_real_)
+  froms <- tibble(
+    cell_id = names(from_mls) %||% character(), 
+    milestone_id = from_mls[.data$cell_id] %>% unname() %>% as.character(), 
+    percentage = from_pct[.data$cell_id] %>% unname() %>% as.numeric()
+  )
+  tos <- progressions %>% select(.data$cell_id, milestone_id = .data$to, .data$percentage)
+  bind_rows(selfs, froms, tos)
+}
+
+#' @rdname convert
+#' @export
+as_list <- function(
+  model,
+  store_dimred = !is.null(model$simulations$dimred),
+  store_cellwise_grn = !is.null(model$experiment$cellwise_grn),
+  store_rna_velocity = !is.null(model$experiment$rna_velocity)
+) {
+  assert_that(
+    !is.null(model$experiment), 
+    msg = "model should be an object that was initialised with `initialise_model()`."
+  )
+  
+  counts <- model$experiment$counts_mrna + model$experiment$counts_premrna
+  progressions <- model$experiment$cell_info %>%
+    select(.data$cell_id, .data$from, .data$to, percentage = .data$time)
+  dataset <- list(
+    id = model$id,
+    cell_ids = rownames(counts),
+    feature_ids = colnames(counts),
+    counts = counts,
+    counts_spliced = model$experiment$counts_mrna,
+    counts_unspliced = model$experiment$counts_premrna,
+    counts_protein = model$experiment$counts_protein,
+    expression = as(log2(counts + 1), "dgCMatrix"),
+    cell_info = model$experiment$cell_info %>% select(-.data$from, -.data$to, -.data$time),
+    feature_info = model$experiment$feature_info,
+    milestone_ids = unique(c(model$gold_standard$network$from, model$gold_standard$network$to)),
+    milestone_network = model$gold_standard$network,
+    progressions = progressions,
+    milestone_percentages = convert_progressions_to_milestone_percentages(progressions)
+  )
+  
+  if (store_dimred) {
+    dimred <- model$simulations$dimred[model$experiment$cell_info$step_ix, , drop = FALSE]
+    rownames(dimred) <- model$experiment$cell_info$cell_id
+    
+    dataset$dimred <- dimred
+    dataset$dimred_segment_points <- model$gold_standard$dimred[!model$gold_standard$meta$burn, , drop = FALSE]
+    dataset$dimred_segment_progressions <- model$gold_standard$meta[!model$gold_standard$meta$burn, , drop = FALSE] %>% 
+      select(.data$from, .data$to, percentage = .data$time)
+  }
+  
+  # add a few more values
+  if (store_cellwise_grn) {
+    regulatory_network <- model$feature_network %>% 
+      select(regulator = .data$from, target = .data$to, .data$strength, .data$effect)
+    regulation_sc <- model$experiment$cellwise_grn
+    
+    regulators <- unique(regulatory_network$regulator)
+    targets <- colnames(dataset$counts)
+    regulatory_network_sc <- 
+      Matrix::summary(regulation_sc) %>% 
+      transmute(
+        cell_id = factor(dataset$cell_ids[.data$i], levels = dataset$cell_ids),
+        regulator = factor(regulatory_network$regulator[.data$j], levels = regulators),
+        target = factor(regulatory_network$target[.data$j], levels = targets),
+        strength = .data$x
+      ) %>% 
+      as_tibble()
+    
+    
+    dataset$regulatory_network <- regulatory_network
+    dataset$regulatory_network_sc <- regulatory_network_sc
+    dataset$regulators <- regulators
+    dataset$targets <- targets
+  }
+  
+  if (store_rna_velocity) {
+    dataset$rna_velocity <- model$experiment$rna_velocity
+  }
+  
+  dataset
+}
+
+conversion_funs <- list(
+  dyno = as_dyno,
+  sce = as_sce,
+  seurat = as_seurat,
+  anndata = as_anndata,
+  list = as_list,
+  none = function(...) NULL
+)
+
+#' @rdname convert
+#' @param format Which output format to use, must be one of 'dyno', 'sce', 'seurat', 'anndata', 'list' or 'none'.
+#' @export
+wrap_dataset <- function(
+  model,
+  format = c("dyno", "sce", "seurat", "anndata", "list", "none"),
+  store_dimred = !is.null(model$simulations$dimred),
+  store_cellwise_grn = !is.null(model$experiment$cellwise_grn),
+  store_rna_velocity = !is.null(model$experiment$rna_velocity)
+) {
+  format <- match.arg(format, choices = c("dyno", "sce", "seurat", "anndata", "list", "none"))
+  fun <- conversion_funs[[format]]
+  
+  fun(
+    model = model, 
+    store_dimred = store_dimred,
+    store_cellwise_grn = store_cellwise_grn,
+    store_rna_velocity = store_rna_velocity
+  )
 }